+++ hera_pspec/__init__.py
@@ -1,13 +1,14 @@
 """
 __init__.py file for hera_pspec
 """
-from hera_pspec import version, conversions, bootstrap, pspecbeam
+from hera_pspec import version, conversions, grouping, pspecbeam
 
 from hera_pspec.uvpspec import UVPSpec
 from hera_pspec.pspecdata import PSpecData
 from hera_pspec.container import PSpecContainer
 from hera_pspec.parameter import PSpecParam
 from hera_pspec.pspecbeam import PSpecBeamUV
+from hera_pspec import uvpspec_utils as uvputils
 
 # XXX: This will eventually be deprecated
 from hera_pspec import legacy_pspec as legacy
