+++ hera_pspec/tests/test_uvpspec.py
@@ -339,23 +339,3 @@ class Test_UVPSpec(unittest.TestCase):
         a = str(self.uvp)
         nt.assert_true(len(a) > 0)
 
-
-def test_conj_blpair_int():
-    conj_blpair = uvpspec._conj_blpair_int(1002003004)
-    nt.assert_equal(conj_blpair, 3004001002)
-
-def test_conj_bl_int():
-    conj_bl = uvpspec._conj_bl_int(1002)
-    nt.assert_equal(conj_bl, 2001)
-
-def test_conj_blpair():
-    blpair = uvpspec._conj_blpair(1002003004, which='first')
-    nt.assert_equal(blpair, 2001003004)
-    blpair = uvpspec._conj_blpair(1002003004, which='second')
-    nt.assert_equal(blpair, 1002004003)
-    blpair = uvpspec._conj_blpair(1002003004, which='both')
-    nt.assert_equal(blpair, 2001004003)
-    nt.assert_raises(ValueError, uvpspec._conj_blpair, 2001003004, which='foo')
-
-
-
